try fixed version

+++ app/api-server.js
@@ -212,48 +212,48 @@ api.delete('/api/picture/:id', api_token_check, function (req, res) {
 	const pictures = db.collection('pictures');
 	// BOLA - API1 Issue here: a user can delete someone's else picture.
 	// Code does not validate who the picture belongs too.
-	pictures.deleteOne({ _id: req.params.id },
-		function (err, result) {
-			if (err) {
-				console.log('>>> Query error...' + err);
-				res.status(500).json({ "message": "system error" });
-			}
-			if (result.deletedCount == 0) {
-				console.log(">>> No picture was deleted")
-				res.status(404).json({ "message": "not found" });
-			}
-			else {
-				console.log('>>> Photo ' + req.params.id + ' was deleted');
-				res.status(200).json({ "message": "success" });
-			}
-		})
-	// pictures.findOne({ _id: req.params.id },
-	// 	function (err, picture) {
+	// pictures.deleteOne({ _id: req.params.id },
+	// 	function (err, result) {
 	// 		if (err) {
 	// 			console.log('>>> Query error...' + err);
 	// 			res.status(500).json({ "message": "system error" });
 	// 		}
-	// 		if (picture && (picture.creator_id == req.user.user_profile._id || req.user.user_profile.is_admin)) { 
-	// 			pictures.deleteOne({ _id: req.params.id },
-	// 				function (err, result) {
-	// 					if (err) {
-	// 						console.log('>>> Query error...' + err);
-	// 						res.status(500).json({ "message": "system error" });
-	// 					}
-	// 					if (result.deletedCount == 0) {
-	// 						console.log(">>> No picture was deleted")
-	// 						res.status(404).json({ "message": "not found" });
-	// 					}
-	// 					else {
-	// 						console.log('>>> Photo ' + req.params.id + ' was deleted');
-	// 						res.status(200).json({ "message": "success" });
-	// 					}
-	// 				})
-	// 		} else {
-	// 			console.log(">>> User does not own the picture")
-	// 			res.status(403).json({ "success": false, "message": "forbidden" });
+	// 		if (result.deletedCount == 0) {
+	// 			console.log(">>> No picture was deleted")
+	// 			res.status(404).json({ "message": "not found" });
+	// 		}
+	// 		else {
+	// 			console.log('>>> Photo ' + req.params.id + ' was deleted');
+	// 			res.status(200).json({ "message": "success" });
 	// 		}
 	// 	})
+	pictures.findOne({ _id: req.params.id },
+		function (err, picture) {
+			if (err) {
+				console.log('>>> Query error...' + err);
+				res.status(500).json({ "message": "system error" });
+			}
+			if (picture && (picture.creator_id == req.user.user_profile._id || req.user.user_profile.is_admin)) { 
+				pictures.deleteOne({ _id: req.params.id },
+					function (err, result) {
+						if (err) {
+							console.log('>>> Query error...' + err);
+							res.status(500).json({ "message": "system error" });
+						}
+						if (result.deletedCount == 0) {
+							console.log(">>> No picture was deleted")
+							res.status(404).json({ "message": "not found" });
+						}
+						else {
+							console.log('>>> Photo ' + req.params.id + ' was deleted');
+							res.status(200).json({ "message": "success" });
+						}
+					})
+			} else {
+				console.log(">>> User does not own the picture")
+				res.status(403).json({ "success": false, "message": "forbidden" });
+			}
+		})
 });
 
 api.delete('/api/admin/user/:id', api_token_check, function (req, res) {
@@ -330,7 +330,7 @@ api.post('/api/picture/file_upload', api_token_check, function (req, res) {
 			return;
 		}
 		
-		var description = random_sentence();
+		// var description = random_sentence();
 		var name = randomWords({ exactly: 2 });
 		name = name.join(' ');
 
@@ -340,7 +340,7 @@ api.post('/api/picture/file_upload', api_token_check, function (req, res) {
 			image_url: imageUrl,
 			name: name,
 			filename: imageName,
-			description: description,
+			// description: description,
 			creator_id: req.user.user_profile._id,
 			money_made: 0,
 			likes: 0,
@@ -383,7 +383,7 @@ api.post('/api/picture/upload', api_token_check, function (req, res) {
 			return;
 		}
 
-		var description = random_sentence();
+		// var description = random_sentence();
 		var name = randomWords({ exactly: 2 });
 		name = name.join(' ');
 
@@ -393,7 +393,7 @@ api.post('/api/picture/upload', api_token_check, function (req, res) {
 			image_url: imageUrl,
 			name: name,
 			filename: imageName,
-			description: description,
+			// description: description,
 			creator_id: req.user.user_profile._id,
 			money_made: 0,
 			likes: 0,
